+++ apps/backend/src/modules/ai-assistant/ai-assistant.controller.ts
@@ -54,11 +54,6 @@ import {
  *           type: string
  *           description: Detailed description of the task
  *           example: "Set a goal to prepare at least two healthy home-cooked meals this week, focusing on incorporating more fruits and vegetables into your diet."
- *         dueDate:
- *           type: string
- *           format: date-time
- *           description: The deadline for the task
- *           example: "2024-09-23T15:15:48.971Z"
  *         label:
  *           type: string
  *           description: A label for the task (priority, tags, etc.)
@@ -312,7 +307,7 @@ class AIAssistantController extends BaseController {
 					}>,
 				),
 			method: "POST",
-			path: AIAssistantApiPath.CHAT_CHANGE_TASK,
+			path: AIAssistantApiPath.CHAT_CHANGE_TASKS,
 			validation: {
 				body: taskActionRequestSchemaValidationSchema,
 			},
@@ -320,14 +315,14 @@ class AIAssistantController extends BaseController {
 
 		this.addRoute({
 			handler: (options) =>
-				this.explainTaskSuggestion(
+				this.explainTasksSuggestion(
 					options as APIHandlerOptions<{
 						body: AIAssistantExplainTaskRequestDto;
 						user: UserDto;
 					}>,
 				),
 			method: "POST",
-			path: AIAssistantApiPath.CHAT_EXPLAIN_TASK,
+			path: AIAssistantApiPath.CHAT_EXPLAIN_TASKS,
 			validation: {
 				body: taskActionRequestSchemaValidationSchema,
 			},
@@ -514,7 +509,7 @@ class AIAssistantController extends BaseController {
 
 	/**
 	 * @swagger
-	 * /assistant/chat/change-task:
+	 * /assistant/chat/change-tasks:
 	 *   post:
 	 *     summary: Change task
 	 *     tags:
@@ -558,16 +553,16 @@ class AIAssistantController extends BaseController {
 		const { body, user } = options;
 
 		return {
-			payload: await this.openAIService.changeTaskSuggestion(user, body),
+			payload: await this.openAIService.changeTasksSuggestion(user, body),
 			status: HTTPCode.OK,
 		};
 	}
 
 	/**
 	 * @swagger
-	 * /assistant/chat/explain-task:
+	 * /assistant/chat/explain-tasks:
 	 *   post:
-	 *     summary: Explain task suggestion
+	 *     summary: Explain tasks suggestion
 	 *     tags:
 	 *       - AI Assistant
 	 *     security:
@@ -600,7 +595,7 @@ class AIAssistantController extends BaseController {
 	 *                       - $ref: '#/components/schemas/ChatMessageTextExplanation'
 	 *                       - $ref: '#/components/schemas/ChatMessageTask'
 	 */
-	private async explainTaskSuggestion(
+	private async explainTasksSuggestion(
 		options: APIHandlerOptions<{
 			body: AIAssistantExplainTaskRequestDto;
 			user: UserDto;
@@ -701,7 +696,7 @@ class AIAssistantController extends BaseController {
 		const { body, user } = options;
 
 		return {
-			payload: await this.openAIService.suggestTasksForCategories(user, body),
+			payload: await this.openAIService.suggestTasksForCategories(body, user),
 			status: HTTPCode.OK,
 		};
 	}