+++ apps/backend/src/modules/ai-assistant/libs/helpers/suggest-task-by-category/generate-suggest-task-response.helper.ts
@@ -3,18 +3,23 @@ import { type z } from "zod";
 import { FIRST_ITEM_INDEX } from "~/libs/constants/constants.js";
 import {
 	AIAssistantMessageValidationSchema,
+	OpenAIErrorMessage,
 	type OpenAIResponseMessage,
 } from "~/libs/modules/open-ai/open-ai.js";
 
-import { ChatMessageAuthor, ChatMessageType } from "../../enums/enums.js";
+import {
+	ChatMessageAuthor,
+	ChatMessageType,
+	HTTPCode,
+} from "../../enums/enums.js";
+import { OpenAIError } from "../../exceptions/exceptions.js";
 import { type ChatMessageCreateDto } from "../../types/types.js";
 import { type taskByCategory } from "./suggest-task-by-category.validation-schema.js";
 
 type TaskByCategoryData = z.infer<typeof taskByCategory>;
 
-const generateTaskSuggestionsResponse = (
+const generateTasksSuggestionsResponse = (
 	aiResponse: OpenAIResponseMessage,
-	taskDeadLine: string,
 ): ChatMessageCreateDto[] | null => {
 	const message = aiResponse.getPaginatedItems().shift();
 
@@ -22,45 +27,49 @@ const generateTaskSuggestionsResponse = (
 		return null;
 	}
 
-	const parsedResult = AIAssistantMessageValidationSchema.safeParse(message);
+	try {
+		const parsedResult = AIAssistantMessageValidationSchema.parse(message);
 
-	if (!parsedResult.success) {
-		return null;
-	}
+		const contentText: string =
+			parsedResult.content[FIRST_ITEM_INDEX].text.value;
+		const resultData: TaskByCategoryData = JSON.parse(
+			contentText,
+		) as TaskByCategoryData;
 
-	const contentText: string =
-		parsedResult.data.content[FIRST_ITEM_INDEX].text.value;
-	const resultData: TaskByCategoryData = JSON.parse(
-		contentText,
-	) as TaskByCategoryData;
-
-	const textMessage: ChatMessageCreateDto = {
-		author: ChatMessageAuthor.ASSISTANT,
-		payload: {
-			text: resultData.message,
-		},
-		threadId: message.thread_id,
-		type: ChatMessageType.TEXT,
-	};
-
-	const taskMessages: ChatMessageCreateDto[] = resultData.tasks.map((task) => {
-		return {
+		const textMessage: ChatMessageCreateDto = {
 			author: ChatMessageAuthor.ASSISTANT,
 			payload: {
-				task: {
-					categoryId: task.categoryId,
-					categoryName: task.categoryName,
-					description: task.description,
-					dueDate: taskDeadLine,
-					label: task.label,
-				},
+				text: resultData.message,
 			},
 			threadId: message.thread_id,
-			type: ChatMessageType.TASK,
+			type: ChatMessageType.TEXT,
 		};
-	});
 
-	return [textMessage, ...taskMessages];
+		const taskMessages: ChatMessageCreateDto[] = resultData.tasks.map(
+			(task) => {
+				return {
+					author: ChatMessageAuthor.ASSISTANT,
+					payload: {
+						task: {
+							categoryId: task.categoryId,
+							categoryName: task.categoryName,
+							description: task.description,
+							label: task.label,
+						},
+					},
+					threadId: message.thread_id,
+					type: ChatMessageType.TASK,
+				};
+			},
+		);
+
+		return [textMessage, ...taskMessages];
+	} catch {
+		throw new OpenAIError({
+			message: OpenAIErrorMessage.WRONG_RESPONSE,
+			status: HTTPCode.INTERNAL_SERVER_ERROR,
+		});
+	}
 };
 
-export { generateTaskSuggestionsResponse };
+export { generateTasksSuggestionsResponse };