+++ apps/frontend/src/pages/chat/libs/components/buttons-controller/libs/components/suggestions-creation-options/libs/enums/suggestions-creation-button-label.enum.ts
@@ -1,5 +1,5 @@
 const SuggestionsCreationButtonLabel = {
-	NO: "No, smth else",
+	NO: "No, something else",
 	YES: "Yes, 3 lowest",
 } as const;