+++ packages/shared/src/modules/ai-assistant/libs/types/ai-assistant-create-multiple-tasks-dto.type.ts
@@ -3,7 +3,7 @@ import { type SaveTextMessageDto } from "./ai-assistant-save-text-message-dto.ty
 
 type AIAssistantCreateMultipleTasksDto = {
 	messages: SaveTextMessageDto[];
-	payload: TaskCreateDto[];
+	tasks: TaskCreateDto[];
 };
 
 export { type AIAssistantCreateMultipleTasksDto };